Added better example for Průběžná Doba form
Průběžná doba uses the jQuery and DataTables CDN libraries. I deleted the previous example because it doesn't show the JSON data fully and also there was an issue in the console that is not related to my App and its function and I wanted a clear Console output in this example picture.